requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.1.31
click==8.1.8
coverage==7.8.0
exceptiongroup==1.2.2
fastapi==0.109.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
pydantic_core==2.33.1
pytest==8.1.1
pytest-cov==4.1.0
sniffio==1.3.1
starlette==0.35.1
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.0
uvicorn==0.27.0

# sqlalchemy
sqlalchemy==1.4.49
databases[sqlite]==0.8.0
